requirements.txt
# Core FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings>=2.0.0
python-multipart>=0.0.6  # Required for FastAPI Form data

# Configuration & Environment
python-dotenv>=1.0.1

# Authentication & Security
bcrypt>=4.0.0  # Password hashing (direct bcrypt, no passlib wrapper)
python-jose[cryptography]>=3.3.0  # JWT token generation/validation

# HTTP Clients & Networking
httpx>=0.24.0
aiohttp>=3.9.0
requests>=2.28.0

# Data Science & Analytics
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.3.0

# Machine Learning & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
joblib>=1.3.0
hmmlearn>=0.3.0

# Trading & Exchange APIs
python-binance>=1.0.19
ccxt>=4.0.0
yfinance>=0.2.28  # Stock/ETF/Bond data

# Task Queue & Background Jobs
redis>=5.0.0
rq>=1.15.0
APScheduler==3.10.4  # In-process periodic task scheduling

# File System & Concurrency
filelock>=3.19.0  # File locking for alert storage
aiofiles>=23.0.0  # Async file I/O for price history cache

# Testing & Browser Automation
selenium>=4.35.0  # E2E browser testing

# Monitoring & Logging (Development)
# Note: These are optional for basic usage
# structlog>=23.0.0  # Structured logging
prometheus-fastapi-instrumentator>=6.1.0
psutil>=5.9.0  # System resource monitoring (CPU, RAM usage)
